requirements.txt
pandas
numpy
python-dotenv
# Añade más dependencias aquí
